fix: don't override pathName property
