+++ test/fixtures/apps/base-context-class/app/router.js
@@ -1,5 +1,3 @@
-'use strict';
-
 module.exports = app => {
   app.get('/', 'home.show');
   app.get('/pathName', 'home.getPathName');
