+++ test/fixtures/apps/base-context-class/config/config.default.js
@@ -1,3 +1,1 @@
-'use strict';
-
 exports.keys = 'test keys';
