+++ test/fixtures/apps/base-context-class/config/config.unittest.js
@@ -1,5 +1,3 @@
-'use strict';
-
 exports.logger = {
   consoleLevel: 'NONE',
 };
